requirements.txt
sanic
aiohttp
requests
gunicorn
gevent
PyYAML
